Allow field-specific custom messages.
Override message with specific text for a particular field. For example instead of "Please lengthen this text to 50 characters or more. ..." you might want "You'll never get  the job with a resume less than 50 characters long."

Works for **all** validation errors. The current behavior of `data-bouncer-message` is preserved - it only effects `patternMismatch` errors. But if you specify a JSON object that looks like the messages array you can override any of the validation messages.

Also made it so you can set `outOfRange` or `wrongLength` to a single string if you want the same message for both edge cases. This works in both `settings.messages` and `data-bouncer-message`.