Add `publicAPIs.showError` and modify `getErrorMessage`
This adds `showError` to the public API, so we can trigger an error on-demand. Using: `bouncer.showError(field, { customMessage: 'Some error.' })`

The second part is modifying `getErrorMessage()` to check to see if we're passing in a `customMessage` object, and to use that for the error, instead of looking at the messages settings.

This is all particularly useful for server-side validation to make use of Bouncer's API's for handling error messages and showing them.